move indexing code to separate file
